web/main/requirements.txt
Flask==0.12
google-cloud-datastore
